Part2 clean

+++ libft_only/ft_calloc.c
@@ -6,20 +6,19 @@
 /*   By: parden <parden@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/30 11:16:12 by parden            #+#    #+#             */
-/*   Updated: 2024/05/21 18:17:14 by parden           ###   ########.fr       */
+/*   Updated: 2024/05/22 16:46:56 by parden           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stdlib.h>
 #include <errno.h>
-#include <limits.h>
 
 void	*ft_calloc(size_t nelem, size_t elsize)
 {
 	void	*ret;
 
-	if (ULONG_MAX / elsize < nelem)
+	if (SIZE_MAX / elsize < nelem)
 	{
 		errno = ENOMEM;
 		return (NULL);
