+++ libft_only/ft_itoa.c
@@ -6,7 +6,7 @@
 /*   By: parden <parden@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 10:40:38 by parden            #+#    #+#             */
-/*   Updated: 2024/05/17 14:54:14 by parden           ###   ########.fr       */
+/*   Updated: 2024/05/22 16:36:36 by parden           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ static size_t	count_char(int n)
 	return (count);
 }
 
-char	*itoa_wrap(long n)
+char	*itoa_wrap(long long n)
 {
 	size_t	len;
 	char	*res;
